Remove upload_path directive from the config file
If the 'upload_path' directive is used, the server will have to find a
way to pass its value to the CGI e.g., by setting an additional
environmental variable. The team agreed that this is outside the scope
of the requirements.

We opt for the simpler alternative, i.e., to set the upload path as the
URI of the location block directly. In other words, the upload path will
be provided to the CGI via the PATH_INFO environmental variable.

For example, the config below will set the upload path to
'/public/folder/'.

location /public/folder/ {
  cgi_pass /cgi/upload.py/
}